Refactor: Openapi관련 api SRP적용
