+++ proma/src/main/java/ai/platform/proma/controller/OpenApiController.java
@@ -2,9 +2,10 @@ package ai.platform.proma.controller;
 
 import ai.platform.proma.domain.User;
 import ai.platform.proma.dto.response.ResponseDto;
-import ai.platform.proma.security.LoginUser;
+import ai.platform.proma.annotation.LoginUser;
 import ai.platform.proma.security.openapi.OpenApiToken;
 import ai.platform.proma.service.OpenApiService;
+import ai.platform.proma.usecase.openapi.OpenApiSignupUseCase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -16,12 +17,13 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/openapi")
 public class OpenApiController {
     private final OpenApiService openApiService;
+    private final OpenApiSignupUseCase openApiSignupUseCase;
 
     @PostMapping("/signup/{promptId}")
     public ResponseDto<OpenApiToken> openApiSignup(
             @LoginUser User user,
             @PathVariable("promptId") Long promptId
             ) {
-        return new ResponseDto<>(openApiService.openApiSignup(user, promptId));
+        return new ResponseDto<>(openApiSignupUseCase.openApiSignup(user, promptId));
     }
 }
