+++ proma/src/main/java/ai/platform/proma/controller/UserController.java
@@ -3,14 +3,11 @@ package ai.platform.proma.controller;
 import ai.platform.proma.domain.User;
 import ai.platform.proma.domain.enums.Role;
 import ai.platform.proma.domain.enums.UserLoginMethod;
-import ai.platform.proma.dto.request.UserRequestDto;
 import ai.platform.proma.dto.response.LoginResponseDto;
 import ai.platform.proma.dto.response.ResponseDto;
-import ai.platform.proma.dto.response.UserResponseDto;
-import ai.platform.proma.security.LoginUser;
+import ai.platform.proma.annotation.LoginUser;
 import ai.platform.proma.service.UserService;
 import jakarta.servlet.http.HttpServletRequest;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
