+++ proma/src/main/java/ai/platform/proma/usecase/openapi/OpenApiSignupUseCase.java
@@ -1,0 +1,8 @@
+package ai.platform.proma.usecase.openapi;
+
+import ai.platform.proma.domain.User;
+import ai.platform.proma.security.openapi.OpenApiToken;
+
+public interface OpenApiSignupUseCase {
+    OpenApiToken openApiSignup(User user, Long promptId);
+}
